+++ app_test.go
@@ -982,35 +982,35 @@ func Test_App_Config(t *testing.T) {
 	require.True(t, app.Config().StrictRouting)
 }
 
-func Test_App_ImmutableString(t *testing.T) {
+func Test_App_SafeString(t *testing.T) {
 	t.Parallel()
 
 	s := "fiber"
 	app := New()
-	same := app.ImmutableString(s)
+	same := app.SafeString(s)
 	if unsafe.StringData(same) != unsafe.StringData(s) { //nolint:gosec // compare pointer addresses
 		t.Errorf("expected original string when immutable is disabled")
 	}
 
 	appImmutable := New(Config{Immutable: true})
-	copied := appImmutable.ImmutableString(s)
+	copied := appImmutable.SafeString(s)
 	if unsafe.StringData(copied) == unsafe.StringData(s) { //nolint:gosec // compare pointer addresses
 		t.Errorf("expected a copy when immutable is enabled")
 	}
 }
 
-func Test_App_ImmutableBytes(t *testing.T) {
+func Test_App_SafeBytes(t *testing.T) {
 	t.Parallel()
 
 	b := []byte("fiber")
 	app := New()
-	same := app.ImmutableBytes(b)
+	same := app.SafeBytes(b)
 	if &same[0] != &b[0] {
 		t.Errorf("expected original slice when immutable is disabled")
 	}
 
 	appImmutable := New(Config{Immutable: true})
-	copied := appImmutable.ImmutableBytes(b)
+	copied := appImmutable.SafeBytes(b)
 	if &copied[0] == &b[0] {
 		t.Errorf("expected a copy when immutable is enabled")
 	}